+++ src/CodeViewer/components/BlockCodeViewer/consts.ts
@@ -1,0 +1,1 @@
+export const SINGLE_LINE_SIZE = 24;
