+++ src/CodeViewer/components/InlineCodeViewer.tsx
@@ -1,0 +1,16 @@
+import * as cn from 'classnames';
+import * as React from 'react';
+
+import { Classes } from '../../classes';
+
+export interface IInlineCodeViewerProps extends React.HTMLAttributes<HTMLElement> {
+  value: string;
+}
+
+export const InlineCodeViewer: React.FC<IInlineCodeViewerProps> = ({ className, value, ...rest }) => {
+  return (
+    <code className={cn(Classes.CODE_EDITOR, className, `${Classes.CODE_EDITOR}--inline`)} {...rest}>
+      {value}
+    </code>
+  );
+};
